Implementation of overlapping character segmentations
This algorithm intents to cut two characters that have a least one pixel between
them. The separation is not a strait line (which is detected by the sliding window
algorithm) but has to be a 'path' from bottom to top. Meaning that you can go
from the bottom of the image to the top without going through a black pixel.
At the end, on the left you'll have one character and on the right the other one